requirements.txt
numpy>=1.26.0
openai>=0.28.1
psycopg2-binary>=2.9.8
PyPDF2>=3.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.1
scikit-learn>=1.3.1
pandas>=2.1.1
matplotlib>=3.8.0